+++ puml-parser-py/buildspec/test_run.bat
@@ -1,0 +1,3 @@
+@echo off
+pumlparser.exe -i "full path to test puml" -o "full path to test json" -t "classdiagram"
+pause
